+++ srv/service.js
@@ -1,10 +1,11 @@
 const cds = require('@sap/cds')
 
-// REVISIT: cds.OutboxService or technique to avoid extending OutboxService
-const OutboxService = require('@sap/cds/libx/_runtime/messaging/Outbox')
+const Base = cds.outboxed ? cds.Service : require('@sap/cds/libx/_runtime/messaging/Outbox')
 
-module.exports = class AuditLogService extends OutboxService {
+module.exports = class AuditLogService extends Base {
   async init() {
+    const outboxed = this.immediate instanceof cds.Service
+
     // add common audit log entry fields
     this.before('*', req => {
       const { tenant, user, timestamp: time } = cds.context
@@ -16,6 +17,10 @@ module.exports = class AuditLogService extends OutboxService {
 
     // add self-explanatory api (await audit.log/logSync(event, data))
     this.log = this.emit
-    this.logSync = this.send
+    // NOTE: logSync is not a public API!
+    this.logSync = (...args) => {
+      if (outboxed) return this.immediate.send(...args)
+      return this.send(...args)
+    }
   }
 }